requirements/pip.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
backcall==0.2.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
coverage==5.2
decorator==4.4.2
Django==3.0
django-extensions==3.0.2
djangorestframework==3.11.0
drf-yasg==1.17.1
freezegun==0.3.15
gunicorn==20.0.4
idna==2.10
inflection==0.5.0
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.2.0
jedi==0.17.1
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.2
more-itertools==8.4.0
packaging==20.4
parso==0.7.0
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.5
psycopg2==2.8.5
ptyprocess==0.6.0
py==1.9.0
Pygments==2.6.1
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
pytest-django==3.9.0
pytest-sugar==0.9.4
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
regex==2020.6.8
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
termcolor==1.1.0
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.5
wrapt==1.12.1
